+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to the "xssencode" extension will be documented in this file
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 
-## [0.0.2] - 2020/12/7
+## [0.0.2] - 2020/12/11
 
 ### Added